[rename] #10 RoomTypeConverter를 Post 도메인으로 변경
