+++ src/main/java/org/gachon/checkmate/domain/post/converter/RoomTypeConverter.java
@@ -1,4 +1,4 @@
-package org.gachon.checkmate.domain.member.converter;
+package org.gachon.checkmate.domain.post.converter;
 
 import jakarta.persistence.Converter;
 import org.gachon.checkmate.domain.member.entity.RoomType;
